adds a DNF page.
Today's weather made me think of my Bighorn, and thinking of Bighorn
made me think of my Bighorn DNF, so I went to UltraSignup to see what
year it was and I saw that it wasn't listed.  It was easy for me to
search through my records to find that it was in 2016, but while I was
doing that I figured I may as well make a brief page that lists all my
DNFs, since when I start back-filling my race reports, I'll want to
start with those.